move functions to test file
